chore: 절대 경로 수정 (#122)
